Run AnalyzerFileReference tests on Core
