+++ src/Compilers/Core/CodeAnalysisTest/Analyzers/AnalyzerConfigTests.cs
