+++ src/Compilers/Core/CodeAnalysisTest/Analyzers/AnalyzerFileReferenceAppDomainTests.cs
@@ -1,0 +1,126 @@
+﻿// Licensed to the .NET Foundation under one or more agreements.
+// The .NET Foundation licenses this file to you under the MIT license.
+// See the LICENSE file in the project root for more information.
+
+#nullable enable
+
+#if NET472 // AppDomains
+
+using System;
+using System.Collections.Immutable;
+using System.Reflection;
+using Microsoft.CodeAnalysis.Diagnostics;
+using Microsoft.CodeAnalysis.Test.Utilities;
+using Roslyn.Test.Utilities;
+using Roslyn.Test.Utilities.Desktop;
+using Xunit;
+
+namespace Microsoft.CodeAnalysis.UnitTests
+{
+    public class RemoteAnalyzerFileReferenceTest : MarshalByRefObject
+    {
+        private Exception _analyzerLoadException;
+
+        public override object InitializeLifetimeService()
+        {
+            return null;
+        }
+
+        public Exception LoadAnalyzer(string analyzerPath)
+        {
+            _analyzerLoadException = null;
+            var analyzerRef = new AnalyzerFileReference(analyzerPath, FromFileLoader.Instance);
+            analyzerRef.AnalyzerLoadFailed += (s, e) => _analyzerLoadException = e.Exception;
+            var builder = ImmutableArray.CreateBuilder<DiagnosticAnalyzer>();
+            analyzerRef.AddAnalyzers(builder, LanguageNames.CSharp);
+            return _analyzerLoadException;
+        }
+    }
+
+    public class AnalyzerFileReferenceAppDomainTests : TestBase
+    {
+        [Fact]
+        public void TestAnalyzerLoading_AppDomain()
+        {
+            var dir = Temp.CreateDirectory();
+            dir.CopyFile(typeof(AppDomainUtils).Assembly.Location);
+            dir.CopyFile(typeof(FromFileLoader).Assembly.Location);
+            var analyzerFile = DesktopTestHelpers.CreateCSharpAnalyzerAssemblyWithTestAnalyzer(dir, "MyAnalyzer");
+            var loadDomain = AppDomainUtils.Create("AnalyzerTestDomain", basePath: dir.Path);
+            try
+            {
+                // Test analyzer load success.
+                var remoteTest = (RemoteAnalyzerFileReferenceTest)loadDomain.CreateInstanceAndUnwrap(typeof(RemoteAnalyzerFileReferenceTest).Assembly.FullName, typeof(RemoteAnalyzerFileReferenceTest).FullName);
+                var exception = remoteTest.LoadAnalyzer(analyzerFile.Path);
+                Assert.Null(exception);
+            }
+            finally
+            {
+                AppDomain.Unload(loadDomain);
+            }
+        }
+
+        [ConditionalFact(typeof(WindowsDesktopOnly), Reason = "https://github.com/mono/mono/issues/10960")]
+        public void TestAnalyzerLoading_Error()
+        {
+            var analyzerSource = @"
+using System;
+using System.Collections.Immutable;
+using Microsoft.CodeAnalysis;
+using Microsoft.CodeAnalysis.Diagnostics;
+using System.Runtime.InteropServices;
+
+[DiagnosticAnalyzer(LanguageNames.CSharp)]
+[StructLayout(LayoutKind.Sequential, Size = 10000000)]
+public class TestAnalyzer : DiagnosticAnalyzer
+{
+    public override ImmutableArray<DiagnosticDescriptor> SupportedDiagnostics { get { throw new NotImplementedException(); } }
+    public override void Initialize(AnalysisContext context) { throw new NotImplementedException(); }
+}";
+
+            var dir = Temp.CreateDirectory();
+
+            dir.CopyFile(typeof(System.Reflection.Metadata.MetadataReader).Assembly.Location);
+            dir.CopyFile(typeof(AppDomainUtils).Assembly.Location);
+            dir.CopyFile(typeof(Memory<>).Assembly.Location);
+            dir.CopyFile(typeof(System.Runtime.CompilerServices.Unsafe).Assembly.Location);
+            var immutable = dir.CopyFile(typeof(ImmutableArray).Assembly.Location);
+            var analyzer = dir.CopyFile(typeof(DiagnosticAnalyzer).Assembly.Location);
+            dir.CopyFile(typeof(FromFileLoader).Assembly.Location);
+
+            var analyzerCompilation = CSharp.CSharpCompilation.Create(
+                "MyAnalyzer",
+                new SyntaxTree[] { CSharp.SyntaxFactory.ParseSyntaxTree(analyzerSource) },
+                new MetadataReference[]
+                {
+                    TestReferences.NetStandard20.NetStandard,
+                    TestReferences.NetStandard20.SystemRuntimeRef,
+                    MetadataReference.CreateFromFile(immutable.Path),
+                    MetadataReference.CreateFromFile(analyzer.Path)
+                },
+                new CSharp.CSharpCompilationOptions(OutputKind.DynamicallyLinkedLibrary));
+
+            var analyzerFile = dir.CreateFile("MyAnalyzer.dll").WriteAllBytes(analyzerCompilation.EmitToArray());
+
+            var loadDomain = AppDomainUtils.Create("AnalyzerTestDomain", basePath: dir.Path);
+            try
+            {
+                // Test analyzer load failure.
+                var remoteTest = (RemoteAnalyzerFileReferenceTest)loadDomain.CreateInstanceAndUnwrap(typeof(RemoteAnalyzerFileReferenceTest).Assembly.FullName, typeof(RemoteAnalyzerFileReferenceTest).FullName);
+                var exception = remoteTest.LoadAnalyzer(analyzerFile.Path);
+                Assert.NotNull(exception as TypeLoadException);
+            }
+            finally
+            {
+                AppDomain.Unload(loadDomain);
+            }
+        }
+
+        private static Assembly OnResolve(object sender, ResolveEventArgs e)
+        {
+            Console.WriteLine($"Resolve in {AppDomain.CurrentDomain.Id} for {e.Name}");
+            return null;
+        }
+    }
+}
+#endif
